requirements.txt
streamlit
duckdb
feedparser
requests
httpx
sentence-transformers
pyyaml
beautifulsoup4
lxml
scikit-learn
langcodes
psycopg2-binary
